fix: split ALB TLS checks and stabilize Kind
